pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aegiscli"
version = "0.2.0"
description = "AegisCLI - Modular Recon Framework"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"

authors = [
    { name = "Martin Simonyan", email = "martinsimonyan2563@gmail.com" }
]

dependencies = [
    "colorama>=0.4",
    "whoisit>=4.0",
    "dnspython>=2.8"
]

[project.scripts]
aegiscli = "aegiscli.cli:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["aegiscli*"]
